web-app/requirements.txt
flask
pymongo

pytest
pytest-flask
coverage

black
pylint
python-dotenv
